POC_agent/requirements.txt
langgraph>=0.2.0
langchain-ollama>=0.1.0
langchain-aws>=0.1.0
guardrails-ai>=0.4.0
pydeid
langsmith>=0.1.0
httpx>=0.27.0
mcp>=1.0.0
pyyaml>=6.0
